flow: move flow code to mt_flow.c
for future rte dev abstract
