+++ lib/src/mt_arp.h
@@ -9,10 +9,10 @@
 
 int mt_arp_parse(struct mtl_main_impl* impl, struct rte_arp_hdr* hdr, enum mtl_port port);
 
-int mt_arp_cni_get_mac(struct mtl_main_impl* impl, struct rte_ether_addr* ea,
-                       enum mtl_port port, uint32_t ip, int timeout_ms);
-
 int mt_arp_init(struct mtl_main_impl* impl);
 int mt_arp_uinit(struct mtl_main_impl* impl);
+
+int mt_arp_get_mac(struct mtl_main_impl* impl, uint8_t dip[MTL_IP_ADDR_LEN],
+                   struct rte_ether_addr* ea, enum mtl_port port, int timeout_ms);
 
 #endif
